requirements.txt
annotated-types
anyio
certifi
colorama
eval_type_backport
griffe
h11
httpcore
httpx
idna
logfire-api
ollama
pydantic
pydantic-ai-slim
pydantic_core
sniffio
typing_extensions
pymongo
dotenv